requirements.txt
numpy 
matplotlib
tensorflow 
gradio
pillow
#-e .